refactor : validate, static 초기화, BigDecimal 변경

+++ src/main/java/com/teamzunior/stringcalculator/operator/OperatorFunction.java
@@ -1,8 +1,10 @@
 package com.teamzunior.stringcalculator.operator;
 
+import java.math.BigDecimal;
+
 @FunctionalInterface
 public interface OperatorFunction {
 
-    double apply(double x, double y);
+    BigDecimal apply(BigDecimal x, BigDecimal y);
 
 }
